Game Update #11
Implemented power up menu which appears when xp bar is full and pauses the game. When player hovers over a powerup, it is highlighted in a red border and the button sound effect plays. When the player clicks on the power up, the menu closes and unpauses the game, and then grants the player the ability selected. If no levels remain of selected ability, the next time the power up menu appears, the option to select that power up will dissapear.